+++ src/main/kotlin/csense/idea/base/bll/kotlin/KtClassOrObject.kt
@@ -1,13 +1,12 @@
 package csense.idea.base.bll.kotlin
 
 import org.jetbrains.kotlin.asJava.classes.KtLightClassForSourceDeclaration
-import org.jetbrains.kotlin.idea.references.resolveMainReferenceToDescriptors
-import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.kotlin.psi.KtNamedFunction
 import org.jetbrains.kotlin.psi.KtProperty
 import org.jetbrains.kotlin.psi.psiUtil.collectDescendantsOfType
+import org.jetbrains.kotlin.psi.psiUtil.isObjectLiteral
 
 fun KtClassOrObject.getProperties() = body?.properties.orEmpty()
 
@@ -43,3 +42,9 @@ val KtClassOrObject.superClass: KtClassOrObject?
     }
 
 fun KtClassOrObject.isCompanion(): Boolean = hasModifier(KtTokens.COMPANION_KEYWORD)
+
+
+fun KtClassOrObject.isUnit(): Boolean =
+    isObjectLiteral() && fqName?.asString() == "kotlin.Unit"
+
+