+++ src/main/kotlin/csense/idea/base/bll/kotlin/KtParameter.kt
@@ -1,9 +1,6 @@
 package csense.idea.base.bll.kotlin
 
 import csense.idea.base.bll.uast.isChildOfSafe
-import org.jetbrains.kotlin.descriptors.ValueDescriptor
-import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptorIfAny
-import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.KtFunctionType
 import org.jetbrains.kotlin.psi.KtNameReferenceExpression